+++ src/main/kotlin/com/hibob/academy/resource/PetsResource.kt
@@ -38,7 +38,7 @@ class PetsResource {
     @PUT
     @Path("/{petId}")
     fun putPet(@PathParam("petId") petId: Long ,updatePet: Pets): Response {
-        val index = allPets.indexOfFirst { p -> p.id.toLong() == petId }
+        val index = allPets.indexOfFirst { p -> p.id == petId }
         if (index >= 0) {
             val petToUpdate = allPets.removeAt(index).copy(id =updatePet.id, dateOfArrival = LocalDate.now())
             return Response.ok(petToUpdate).build()
@@ -50,7 +50,7 @@ class PetsResource {
     @DELETE
     @Path("/{PetId}")
     fun deletePet(@PathParam("PetId") petId: Long): Response {
-        val index = allPets.indexOfFirst { p -> p.id.toLong() == petId }
+        val index = allPets.indexOfFirst { p -> p.id == petId }
         if (index >= 0) {
             allPets.removeAt(index)
             return Response.status(Response.Status.OK).entity("Pet deleted").build()
